refactorización para acomodar los archivos y que se vieran más bonitos, muy orgullosa la verdad
