+++ compilador/lexer.py
@@ -1,4 +1,3 @@
-import ply.lex as lex
 import sys
 
 reserved = {
