+++ compilador/tests/test_3.txt
