+++ compilador/tests/test_6.txt
