+++ compilador/tests/test_8.txt
